src/use-cases/orgs/authenticate.ts
import { Org } from "@prisma/client";
import { OrgsRepository } from "../../repositories/orgs-repository";
import { compare } from "bcryptjs";
import { InvalidCredentialError } from "../errors/invalid-credential-error";

interface AuthenticateUseCaseRequest {
  email: string
  password: string
}

interface AuthenticateUseCaseResponse {
  org: Org
}

export class AuthenticateUseCase {
  constructor(private orgsRepository: OrgsRepository) {}

  async execute({
    email,
    password,
  }: AuthenticateUseCaseRequest): Promise<AuthenticateUseCaseResponse> {
    const org = await this.orgsRepository.findByEmail(email)

    if (!org) {
      throw new InvalidCredentialError()
    }

    const doesPasswordMatches = await compare(password, org.password_hash)

    if (!doesPasswordMatches) {
      throw new InvalidCredentialError()
    }

    return {
      org,
    };
  }
}
